refactor(segment): Moved persistense directory into infrastructure
