+++ module/segment/src/Infrastructure/Persistence/Projector/Segment/DbalSegmentCreatedEventProjector.php
@@ -7,16 +7,17 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Segment\Persistence\Dbal\Projector\Segment;
+namespace Ergonode\Segment\Infrastructure\Persistence\Projector\Segment;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
-use Ergonode\Segment\Domain\Event\SegmentNameChangedEvent;
+use Ergonode\Segment\Domain\Event\SegmentCreatedEvent;
+use Ergonode\Segment\Domain\ValueObject\SegmentStatus;
 use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class SegmentNameChangedEventProjector
+class DbalSegmentCreatedEventProjector
 {
     private const TABLE = 'segment';
 
@@ -41,19 +42,21 @@ class SegmentNameChangedEventProjector
     }
 
     /**
-     * @param SegmentNameChangedEvent $event
+     * @param SegmentCreatedEvent $event
      *
      * @throws DBALException
      */
-    public function __invoke(SegmentNameChangedEvent $event): void
+    public function __invoke(SegmentCreatedEvent $event): void
     {
-        $this->connection->update(
+        $this->connection->insert(
             self::TABLE,
             [
-                'name' => $this->serializer->serialize($event->getTo(), 'json'),
-            ],
-            [
                 'id' => $event->getAggregateId()->getValue(),
+                'code' => $event->getCode(),
+                'name' => $this->serializer->serialize($event->getName(), 'json'),
+                'description' => $this->serializer->serialize($event->getDescription(), 'json'),
+                'status' => SegmentStatus::NEW,
+                'condition_set_id' => $event->getConditionSetId() ? $event->getConditionSetId()->getValue() : null,
             ]
         );
     }
