+++ module/segment/src/Infrastructure/Persistence/Query/DbalSegmentProductsQuery.php
@@ -7,14 +7,13 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Segment\Persistence\Dbal\Query;
+namespace Ergonode\Segment\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
 use Ergonode\Grid\DbalDataSet;
 use Ergonode\Segment\Domain\Query\SegmentProductsQueryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\SegmentId;
-use Ergonode\SharedKernel\Domain\Aggregate\ProductId;
 
 /**
  */
